restructure redux folder
