+++ frontend/src/store/index.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "./reducers/auth.reducer";
+import authReducer from "./auth/slice";
 
 export default configureStore({
     reducer: {
